Skip LOGNAME dependent test if LOGNAME not defined
Most non-Windows execution environments define the LOGNAME environment
variable.  When running the Jenkins Dockerfile-slim, it does not appear
to define that environment variable on the master.  Rather than create
complicated searching for another environment variable whose value is
valid as a branch name, we will skip this test if LOGNAME is not set.

LOGNAME is set in enough contexts (like ci.jenkins.io and developer
desktop machines and other test environments) that it is not a good use of
time to embed additional logic into the test for a relatively rare case.
Skip the test if the environment won't support the test.