[JENKINS-19022] add extension to stop storing buildsByBranch history
We store a new copy of the entire history buildsByBranch every build, as
part of BuildData. This is used for a few cases of determining which
revisions already built, and how to select the branch. Unfortunately, it
bloats the build.xml and takes time to process, as it grows indefinitely
under some use-cases, such as using the GerritTrigger plugin, in which
every build is essentially run on a different branch.

To avoid this, add an extension which allows disabling the saving of
this buildsByBranch history, only maintaining the current data for each
build.

Add tests to make sure that buildsByBranchName tracks history by
default, and doesn't track history when the extension is enabled.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>