Fix SECURITY-723 - info disclosure over http
Prior to this change, the /git/search/ URL responded to queries for
user name information with the user names of system users which match
the query parameters, even if the query is not authenticated.

This removes the ability to search through the /git/search/ URL.  The
/git/search/ URL is an accident of implementation from the 2011
addition of the GitStatus object.