Suppress findbugs warnings from Jenkins 2.60 builds
Some tests call methods that rely on a Jenkins instance but the test is
being called outside a JenkinsRule.  Thus there is not a valid Jenkins
instance.  Rather than alter behavior of those tests, suppress the
findbugs warning for the Jenkins.getInstance() reference.

Problem is visible with:

    mvn clean -Djenkins.version=2.60.1 -Dtest=StashTest test findbugs:findbugs

These exclusions and the null checks should be removable once the JDK
8 pull request is merged.