Fetch tags from public repo for GitSCMFileSystemTest
Tests running in a continuous integration job from a private forked repo
may fail because the job itself does not have access to the private fork
during its execution.  The checkout step of the job uses the credentials
for the private forked repo, but (correctly) does not allow the job
access to those credentials.

Found this issue while fixing SECURITY-723 Unauthenticated user
information disclosure

The fetch is only performed if the required tags are not already found
in the repository.