[JENKINS-19022] rebuild BuildData objects instead of storing them in the build.xml
The BuildData objects are serialized into the build.xml for every build.
This is normally not a problem, as their contents are small. However,
for certain types of builds (gerrit-triggered projects, really any git
build which uses a large number of refs) the BuildData objects can
become enormous. Worse, in those cases the data it saves is nearly
worthless.

Unfortunately, complete removal of the BuildData objects is not
possible, as we rely on this data to correctly predict change logs, and
to determine what revisions to build. Disentangling all of these
operations from the BuildData is tricky, if not impossible. This is
especially so for extension points such as the BuildChoosers.

Instead, lets opt for a different approach: generate the BuildData
instead of serializing the buildsByBranchName map to the build.xml for
every single build.

Do this by implementing a simpler BuildDetails structure, which contains
most of the same information, but does not contain the
buildsByBranchName map.

Stop storing the BuildData object into the build Actions. Instead,
search for BuildDetails and regenerate the BuildData structure from
this. As a fallback, allow finding old BuildData actions and using
those, if they still exist.

Jenkins Administrators may wish to purge this data from their build.xml
files if they do not need the historical data, but for now, keep the
plugin as backwards compatible as possible.

With this change, we now only store a single Build object per build
instead of storing the entire branch map, which should significantly
reduce build.xml bloat.

If the time cost of rebuilding the BuildData structure is too high,
(such as when there are a large number of historical builds), we could
add some sort of limit on the number of past builds to check. However,
this would require some thought, as it would not produce exact replicas
of the current BuildData objects.

See [JENKINS-19022] and  [JENKINS-47789] for more details on the issues
the BuildData objects can cause.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>