fix JENKINS-50556 local branch names cause polling to return true
When the git plugin attempts to determine if there are new changes using
the DefaultBuildChooser it can sometimes accidentally report changes
even when there are no new changes.

This can occur in rare cases due to the way the code handles non-branch
references.

Currently, we first check the reference pre-fixed with the remote name,
assuming that its a branch. getHeadRevision is used to find all the
proper revisions. However, when we are polling, we exclude any already
built changes. Since in this scenario, the remote hasn't updated, we'll
report that no changes exist.

Because we didn't find any valid references, we then check for whether
the reference is actually a tag or other specific value (such as
"refs/changes/<change number>" from gerrit, by checking the ref
directly.

In some cases, it is possible for the local work tree to have a checkout
of the local branch which is at some commit that maybe hasn't been built
yet. (For example, if the job in question creates commits or performs
git commands).

This results in the polling reporting that a new commit exists, which
triggers a build. However, the actual build will not be for that commit,
so we end up in an endless loop where polling reports a new commit every
time.

The issue is somewhat rare for a number of reasons:

 (a) it requires a local branch to be checked out, which does not happen
     by default when checking out using the git plugin
 (b) the local branch must be checked out to a commit that is *not* on
     the remote

This is somewhat unlikely but possible if a job happens to generate new
commits, or perform other work as part of its tasks. Once the branch is
created, it causes the endless rebuilding.

The real root cause is due to how we handle scrubbing already built
revisions. We currently discard them directly in getHeadRevision.

Instead, move this check outside of getHeadRevision only *after* we
check for tags. This ensures, that the check for non-branch refs will
only occur in the case where there *are* no matching remote names. This
avoids accidentally looking at the local branch name in this case,
ensuring that we do not report its commits as candidates for a build.

This ensures that we'll match commit selection when polling vs when not
polling, and prevent the endless rebuild cycle.

Add a test case highlighting the issue so that we can prevent
regressions in the future.

Fixes: JENKINS-50556
Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>