Use addAll instead of replacing the revisions collection
A future patch is going to migrate the code which ignores already built
elements from outside of getHeadRevision directly into the
getCandidateRevisions.

However, currently when we check for tag-like refs we just assign
revisions to the collection returned by the getHeadRevision. This will
cause a problem because the Collections.singleton() is immutable and does
not support removing elements.

Fix this by using .addAll to put the elements into the HashSet rather
than continuing to use the singleton collection.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Title: Use addAll instead of replacing the revisions collection
Change-type: DefectResolution