using author in the changelog should not require workspace polling
Commit d8338ce4bca9 ("[FIXED JENKINS-19001] extension can declare
requirement for a workspace during polling", 2014-03-06) appears to have
accidentally set the workspace to poll whenever AuthorInChangelog is
used.

The extension only modifies two things, (a) the changelog parser which
should not depend on polling and (b) the check for user exclusion in
isRevExcluded.

One might consider that the second case does depend on polling, however
it is only triggered in the case where the UserExclude extension is
enabled! Thus, we should not need to poll when only using the
AuthorInChangelog extension.

It appears this functionality should already be tested by
testAuthorOrComitterTrue() in GitSCMTest.java.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>