don't require polling for ChangeLogToBranch
This extension only modifies how we compute the changelog which does not
occur during polling, and thus should not impact how polling behaves.
The extension does mark that it requires the workspace when polling, but
this seems incorrect. Behavior should be identical regardless of whether
we poll in the workspace or not.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>