[SECURITY-810] require ADMINISTER for repo browser check
I've confirmed with a test script in my jenkins-bugs-private directory
that prior to this change, these 5 classes opened the external URL that
was passed as an argument in a GET request and returned a message even
if the user was not authenticated.

I've confirmed with this change that they now behave the same as the
FisheyeGitRepositoryBrowser referenced in SECURITY-810.  If they receive
a GET request from a user without ADMINISTER permission, they return
"<div/>" and do not open the external URL.