Include experimental git client in ATH
This git client requires an experimental git client because the git
client plugin is not ready to release without being verified with the
matching git plugin release.  The git client plugin experimental release
is available in the experimental update center, but the ATH framework
(correctly) does not load plugins from the experimental update center.

Copying the git-client hpi file to the target directory causes the ATH to
load the beta git client plugin.