[JENKINS-53050] Fix absent value handling for shallow depth parameter
A reference instead of primitive type has to be used. Otherwise when
in the web form the depth field is left empty, this has to be converted
to an int primitive value. The result is 0 which is persisted in the
configuration and then shown the next time in the form.

The initial value of 1 is not required anymore, null is fine and even
better describes the absent value. On execution the value has anyway to
be checked against absent and invalid values.