Extend GitTagActionTest for more cases
Tests are imperfect but better than no tests.  The tests are able to
run a Freestyle project and create two different tags in the workspace
of the project.

The tests are not able to create the tags from the doSubmit method and
thus are still unable to assert several important cases.