Catch and print MissingObjectException
This exception can occur if the repository JGit sees does, for whatever reason, not correspond to the repository Git sees.
For instance, JGit does not support git-worktree, so it will report that it could not find the object Jenkins was looking for.

This change allows git-plugin to work with git-worktree again.

Fixes JENKINS-53725