Confirm commit summary truncates with CLI git
JENKINS-29977 reports that command line git truncates on the first
word boundary preceding character 73 in the commit summary.

JGit does not truncate.

This test is intended to confirm that the existing behavior is
retained.  New optional behavior is being prepared based on a global
switch that will disable truncation of long commit summaries.