Set expected value in test, not in data generator
The truncated summary test for JENKINS-29977 is easier to read
and easier to extend if the expected value is computed in the
test method rather than in the data generator method.

The JENKINS-29977 fix moves the truncation decision from inside
the git client plugin to inside the git plugin.  That change is
easier to represent in the test method than in the data generator
method.