JENKINS-29977 add option to retain full changelog summary
JENKINS-29977 notes that the command line git implementation inside git
client plugin 2.0 and later will silently truncate the commit summary
at the last word break prior to 73 characters.

The JGit implementation does not truncate the commit summary.

This change adds a global option which will retain the full commit summary
in both the command line git implementation and the JGit implementation.

The command line git implementation continues to truncate by default
so that behavioral compatibility is retained for existing users of the
command line git implementation.

The JGit implementation continues to not truncate by default in the
user interface so that behavioral compatibility is retained for existing
users of the JGit implementation.

Consumers of the git plugin API that use the previous two argument
GitChangeSet constructor with the command line git implementation will
continue to see truncated commit summary messages unless the global
option has been set to retain the full commit summary.

Consumers of the git plugin API that use the previous two argument
GitChangeSet constructor with the JGit implementation will see truncated
commit summary messages unless the global option has been set to retain
the full commit summary.  This is a change of behavior from previous use.
If an API consumer needs the previous behavior, they will need to call the
new three argument GitChangeSet constructor so that they can explicitly
control the truncation.

This change relies on the git client plugin 3.0 change which moves the
responsibility for commit message summary truncation from the git client
plugin to the git plugin.