Remove getBySHA1 REST API obsoleted by SECURITY-595
The SECURITY-595 security fix for Jenkins core prevents access to the URL
/job/<Job Name here>/scm/. That prevents access to …/scm/bySHA1/….
That change breaks this undocumented feature with no test coverage.

Since Git plugin 4.0 is a major release, let's use the major release to
remove this unusable API.

If we find a critical need to provide this REST API, it can be moved to
the Descriptor.  Descriptors are already involved in URL handling and
will work out of the box.