Skip testGitSCMCanBuildAgainstTags if CLI git 2.20+
Git client plugin 2.7.6 and later include a fix for the tag handling
behavior change of command line git 2.20.  However, we don't want to
force users to upgrade to that git client version, so the git plugin
depends on an older version.

The older git client version does not include the CLI git 2.20 tag
handling change so it will fail this test.  Rather than fail the test,
we exit the test early if it is running with CLI git 2.20 or newer.