[JENKINS-57493] Fix unit test failure in DE
Remove assertion that depends on the locale of the command line git
process.  No significant value checking for a specific English language
message from command line git.  Command line git is increasingly localized
and will continue to be better and better localized.  Testing for a
specific English message doesn't help enough for the distraction it
creates due to failing tests.  Later checks in the test confirm that
the failure happened.