[JENKINS-56176] Tests for token macro expansion
When earlier changes stopped setting BuildData, the token macro
replacement also stopped working because it is dependent on BuildData
contents.  No tests confirmed that behavior so there was no warning that
the functionality was lost.

These tests fail on the master branch and pass with the changes that
revert the BuildData changes.