[JENKINS-42609] Improve env var descriptions
The environment variables available from the git plugin have surprising
differences related to how the values of the variables are calculated.
Many of the variables are calculated based on the checkout performed by
the current job in the current workspace.  Those include:

 * GIT_BRANCH
 * GIT_COMMIT
 * GIT_LOCAL_BRANCH
 * GIT_PREVIOUS_COMMIT
 * GIT_PREVIOUS_SUCCESSFUL_COMMIT
 * ...

The values of four of the variables are calculated based on the git
global configuration from 'Configure System' of the 'Manage Jenkins' page.
They are:

 * GIT_AUTHOR_EMAIL
 * GIT_AUTHOR_NAME
 * GIT_COMMITTER_EMAIL
 * GIT_COMMITTER_NAME

My incorrect assumption had been that the values of those four variables
were calculated based on the most recent commit in the current workspace.
That is not correct.

They are taken from the global configuration in 'Configure System'
of the 'Manage Jenkins' page.