[JENKINS-58964] Test git is considered a "Modern SCM"
Test that the git plugin provides a modern SCM implementation.

API extensions added in workflow-cps-global-lib 2.14 are used in git
plugin 3.11.0 and later and are required for it to be recognized as a
modern SCM implementation.

The git plugin must require workflow-cps-global-lib 2.14 as its minimum
version in order to be recognized as a modern SCM implementation and
still fix JENKINS-43802.