[Fix JENKINS-58964] Require Jenkins 2.138.4 minimum
The workflow-global-cps-lib 2.14 API addition to fix JENKINS-43802 needs
to be a minimum requirement for the git plugin, otherwise the git plugin
is not recognized as a modern SCM implementation.

Release 2.14 of workflow-global-cps-lib requires Jenkins 2.138.4 as its
minimum version. Git plugin must require Jenkins 2.138.4 as its minimum
version as well.

Increment the plugin version number to 3.12.0 due to that increase in
minimum Jenkins version.