Reduce test configurations by 50%
Have not seen a failure on Windows specific to Java 11.  Since Windows
tests run slower and I have fewer Windows machines available, reduce
the Windows cases from 3 to 1, testing Java 8 with the default Jenkins
(2.138.1).

Have not seen failures on Linux with the default Jenkins that are not
also visible on Windows with default Jenkins.  Test Linux with Java 8
and Java 11 using the recent LTS release.