Remove MethodUtils - use Util class
Now that the base Jenkins version is 2.138.1, no need for the previous
implementation of MethodUtils.

Jenkins 1.633 and later includes the Util class changes which replace
MethodUtils. Confirmed with the code coverage report that the tests cover
the same lines after the replacement of MethodUtils with the hudson.Util
class and its curent implementation of isOverridden.