[JENKINS-57694] Fix calling SpecificRevisionBuildChooser#getDisplayName()
There may be `BuildChooser`s that don't have a `Descriptor` because
they are not meant to be selectable by a user in the UI. Currently
`SpecificRevisionBuildChooser` is such a `BuildChooser`.

This missing `Descriptor` was causing an exception when
`BuildChooser#getDisplayName()` was called within
`GitSCM#compareRemoteRevisionWithImpl()` to log its name.

The `Descriptor` was once already in place but then removed in commit 75e78800
because the `SpecificRevisionBuildChooser` is not meant to be selectable by
a user in the UI.

Therefore the absence of a `Descriptor` has to be handled.