Require plugin versions from BOM
Switches from internally maintaining a compilable version number for
each component to specifically use the version provided by Jenkins
plugin BOM 2.138.2.

As more and more plugins adopt the plugin BOM, users will naturally be
running the same plugin versions that are used to test the git plugin
interactively and with automation.

Strategy change is inspired by a desire to move users onto the most
heavily tested plugin versions.  Users that don't want to update
dependencies are welcome to choose to remain with older versions of
this plugin.