[JENKINS-42860] Whitelist pipeline safe getters
Pipelines are not allowed to reference plugin fields unless those fields
are whitelisted.  Allow plugin access to many fields in the git plugin.

Test automation of this change uses an acceptance test rather than a test
included inside this repository. There are suggestions from the community
that test automation of a whitelisted method should be feasible, but I
didn't want to delay the release of git plugin 4.1.0 for that automation
inside the plugin when there is already automation outside the plugin.

See the acceptance test docker image at:
https://github.com/MarkEWaite/docker-lfs/tree/lts-with-plugins

See the acceptance test source code at:
https://github.com/MarkEWaite/jenkins-bugs/blob/JENKINS-42860/Jenkinsfile