Restrict test to fork a single process
Setting forkCount to `1C` causes the tests to hang completely on systems
with a large number of available CPU cores.  Rather than risk being
unable to test on a full range of computers, limit the forkCount to 1.

Eventually it would be best to identify the root causes of the hanging
tests, but that's far beyond what I can do now.