[JENKINS-60617] Include examples in the git step help
This file is displayed as help for the git step when using the
pipeline syntax generator.

It is also displayed at https://jenkins.io/doc/pipeline/steps/git/.
That page has received the most negative reviews of any page on
https://jenkins.io/ .  The most frequently requested enhancement for
the page is to include examples.  This change includes examples.

A high frequency complaint is that the git step is missing
functionality.  That will not change.

The checkout step is the replacement for the git step.  The checkout
step provides all the functionality of the git step and all the
functionality available from the git plugin. It is more maintainable
and better prepared for use in realistic pipelines.

This page also includes strong guidance to use the checkout step.