More separation between git step examples and args
The formatting of https://jenkins.io/doc/pipeline/steps/git/ drops the
hard returns ( '<hr>' ) that are used to separate the examples from the
arguments.

The hard returns are not dropped from the version of the git step help
visible inside Jenkins at:
http://jenkins-url/pipeline-syntax/html#git-step-with-git-and-polling

As a compromise, place bold text 'Argument Descriptions' immediately after
the last example. Ends the examples and makes the online generated version
at https://jenkins.io/doc/pipeline/steps/git/ a little easier to read.

The change to https://jenkins.io/doc/pipeline/steps/git/ won't be visible
until the official release of the git plugin includes this commit.
That will be a release *after* git plugin 4.1.0.  This change is *not*
included in git plugin 4.1.0.