Exclude jsoup from git plugin jpi packaging
The jsoup jar file is added to the hpi file as a transitive dependency of
the configuration as code test harness.  The git plugin does not deliver
the configuration as code test harness and so it should not deliver the
dependencies of that test harness either.

We've had issues in the past where other plugins mistakenly depended on
a jar file that the git plugin bundled (like joda-time). Better to avoid
others from depending on a transitive test dependency by not including
the transitive test dependency.

The jsoup jar file was included in git plugin 4.1.0 and 4.1.1.
Earlier versions did not include it.

https://github.com/jenkinsci/git-plugin/pull/814/commits/f3c65cb72548d3ab8ed863d5817624119b199455
added the transitive dependency.