Remove mocks from crumb exclusion test
Since the test already had a JenkinsRule, use it to make passing
and failing HTTP requests to the notifyCommit URL and to other
URL's.

Uses a single JenkinsRule for the entire class of tests, since
none of them alter the state of Jenkins enough to justify
restarting Jenkins.  Saves the time of stopping and starting
Jenkins for each test.