Prevent test failure due to busy Windows log file
Windows job cleanup fails to delete build logs in some of these tests.
Wait for the jobs to complete before exiting the test so that the
build logs will not be active when the cleanup process tries to
delete them.

The build logs which are not being promptly released are logs that
Jenkins core maintains.  It seems like Jenkins core should be the
one that is responsible to assure the locks are released before the
test exits.  However, that might slow tests by blocking their exit
unnecessarily on platforms that allow open files to be deleted.