Clarify that tag names are not supported by git step
The git step is a simplified, intentionally less capable
implementation.  It will not be extended any further.  It is retained
for compatibility, but will not be used for any future additions.

The checkout step is the replacement for the git step.  The checkout
step provides all the functionality of the git step and much more.

User feedback on the git step page was:

  is specifying a tag name for the git branch setting still supported?
  It doesn't seem to work any more, so maybe this doc needs updating?"

The doc already states that tag names are not supported.  The reader
did not understand that non-support from the existing text.  This pull
request attempts to make the non-support clearer by formatting the
non-supported items as a list (for emphasis) and by adding tag names
to the list rather than relying on the earlier list that included tab
names as not supported in the git step.