Add images to improve docs clarity
The images provide more context for the reader.  They help the reader
confirm that the screen they are seeing in their running Jenkins instance
is similar to the screen being documented.

Accepts that the conversion from tables to divs will justify revising
all these screenshots, but that is not likely to arrive in an LTS for
at least 3 months.