Enable CleanBeforeCheckout to be decorated in the CloneCommand The concern of removing the second fetch call is a possible repository data loss or misconfiguration in terms of the extra behaviors applied during checkout.
If the second fetch call is ignored, CleanBeforeCheckout option will also be ignored
as it doesn't implement the decorateCloneCommand which is used by the git fetch call.
To ensure that it is not ignored, decorateCloneCommand has been implemented for this
particular extension. The unit test GitSCMTest#testCleanBeforeCheckout doesn't fail
after the removal of second fetch call now.