[JENKINS-57587] Prevent GitSCMFileSystem NPE on '*' branch name
This has the questionable result that by avoiding the null pointer
exception it now allows a Pipeline job that is not a multibranch
Pipeline to switch from one branch to another inside the same job.

The technique of using a single job to validate multiple branches was
commonly used with Freestyle jobs because they did not have the
concept of multibranch.  It is generally much better to allow Jenkins
to manage the Pipeline jobs with multibranch rather than having a
single job that switches between branches.  However, a null pointer
exception is not a friendly way to tell users that they should not
use a specific technique.