Depend on parameterized trigger plugin 2.35.1
Do not force parameterized trigger plugin users to update to a newer
version than required.  Majority of users are running 2.35.2 or later.
Even more likely to be running 2.35.2 or newer if they are running
Jenkins 2.204.1 or later as required by git client plugin 3.3.0.

If the user installs git client plugin 3.3.0, they must be running Jenkins
2.204.1 or newer.  Requiring parameterized trigger plugin 2.35.1 or newer
does not change the parameterized trigger plugin version installed on all
but a very few installations.

91% of all parameterized trigger plugin installations are running 2.35.1
or newer.  That's enough to justify this increase in the minimum required
parameterized trigger plugin version in the git plugin.