Use correct branch in testAvoidRedundantFetchWithoutHonorRefSpec
The previous implementation of the test would create a new CheckoutCommand
with git.checkout() but then did nothing with it.  It created the
"foo" branch but did not checkout the "foo" branch.  The commit that
was described as "Commit in foo" was a commit to the master branch,
not a commit to the foo branch.