Fix testAvoidRedundantFetchWithHonorRefSpec assertions
The previous implementation of the test failed when the branch name
'master' is used because the Jenkins job directory name includes the
branch name.  This now asserts more precisely on the content of the
FETCH_HEAD file.

As part of the exploration to find the problem, additional assertions
have been added to confirm that the correct SHA1 is mentioned in the
FETCH_HEAD file.

Use correct branch in GitSCMTest.testAvoidRedundantFetchWithHonorRefSpec

The previous implementation of the test would create a new CheckoutCommand
with git.checkout() but then did nothing with it.  It created the
"foo" branch but did not checkout the "foo" branch.  The commit that
was described as "Commit in foo" was a commit to the master branch,
not a commit to the foo branch.