[JENKINS-37263] prefere remote reference for checkout
If a branch that contains a slash is configured as branch to build and
local branch behaviour is configured, then two candidate revisions will
be found in case remote branch is ahead of local: configured local
branch name and branch name with prefix 'refs/remotes/<remote>/'. The
first result will be taken, which can be a local or remote branch
depending on sha1 of the revisions.