[change in design] Make better decision of recommending git tool with adding exisiting git tool provided by system
Currently, a git client is initialised with a git implementation depending on two decisions:
	- User: A user can choose between git, jgit and jgitapache and if not chosen, the default
		implementation is "Default" (git).
	- System: The system then takes the user's choice and validates if the chosen implementation is
		installed within the system.
Now we have another decision variable for this process:
	- GitToolChooser: It takes the tool.getGitExe() from the system's recommendation and tries to suggest
			an optimal implementation on the basis of size of remote repo.

Now, the problem in this process was that GitToolChooser was not taking the system's suggestion to recommend the optimal
git tool. This would break some existing use cases.

Solution: This commit introduces the gitExe as an addition argument to the GitToolChooser which will then take this information
into consideration to provide the optimal git implementation without breaking existing use cases.

Note: Please look at test class to understand different use cases.