Add a new extension decoration method
This method can be called by an extension to convey if a certain extension
feature is supported by JGit or not.

After this information is provided by the an extenions or multiple extensions, the
command can determine if JGit is to be used or not.

Even if one extension is not supporting a feature, JGit will not be recommended.