[JENKINS-63539] Use additional repo URL variants to find cache
Estimation of repository size uses a cache key based on
the repository URL.  There are several different forms of
repository URL that can represent the same git repository.

For example, the following URL's all point to the same repository

* git://github.com/jenkinsci/git-plugin
* git@github.com:jenkinsci/git-plugin.git
* https://github.com/jenkinsci/git-plugin
* https://github.com/jenkinsci/git-plugin.git
* ssh://git@github.com/jenkinsci/git-plugin.git

This pull request permutes the repository URL provided by the user into
several other forms in hopes that one of the forms already has a cached
copy that can be used for local estimation of repository size.  This
assumes that local Java processing of strings is much faster than a REST
API call to the hosting provider to request the size of the repository.

Includes tests to confirm remote alternatives

The alternatives work well for repository URL formats for GitHub, GitLab,
Gitea, and most other providers.  The alternatives do not work well with
Bitbucket URL formats because Bitbucket applies the user name differently
than the other providers.

Use random values in locations where the value of the random will have
no affect on the assertions.

Do not return null from remote alternatives

Cleaner interface if it always returns a value