[JENKINS-63541] Don't create empty dirs in chooser
The git plugin previously created an empty directory in the
$JENKINS_HOME/caches directory for any repository URL that it checked for
size. Those empty directories are not likely to use too much disc space,
but a large installation that clones many different repositories will
have many empty directories in the caches directory.

Provides a new getCacheDir API that returns null if the directory does
not already exist.