Introducing recommendGitToolOnAgent In reference to the bug JENKINS-63519 which exposed an issue where the GitToolChooser is not able to determine the correct implementation if a node-specific installation is present in the system.
This API is called when the system wants to recommend `cli git` and the user has chosen `jgit`.
To find the right cli git, it iterates over the available installations, checks if one of them
contains a property map, confirms with the node attributes (using GitUtils.resolveGitTool) and then
finally suggests the right git implementation.

With this change, we have added two new parameters: Node path and TaskListener. Both of them are needed
for better estimation of the correct git tool present on a particular agent.