Add more tests to describe the behavior of GitToolChooser if user's choice is JGit and the size of repo> 5M
For the GitToolChooser, the most difficult scenario is when the user has chosen
jgit and size of repo > 5M.
In this case, the system has to decide which command line git implementation will
be the right implementation for that particular environment without breaking any possible
use-case.

With the added test cases what we want to say is (keeping the above mentioned conditions same):
- if there is a node specific installation present, that will be suggested after being verified that
it is actually installed in that environment.
- if there is no node specific installation, the default implementation will be suggested.
There is one side-effect of this implementation, if the only implementation is `jgit` in the system
then that will be recommended instead of command line git. But that wouldn't affect the user since there
would be no possible change in the git implementation used for the git client.