[JENKINS-64019] Rephrase repository help
Include the "path" phrase specifically since the Jenkins tutorials use
that technique and others may use the technique when they are not
using agents to perform builds.

Using a path is a mistake when agents are performing the build, since
Git is already well equipped to transmit changes efficiently over the
network.  Using a path to access a remote file system is likely to be
much less efficient than using the remote communications protocols
that are already part of git.