[JENKINS-64026] Only add git tag action if administrator enabled it
The git tag action attempts to apply a tag in a Jenkins workspace from
a completed job.  It does not perform any other action on the tag except
attempting to create it in the workspace.

If the workspace does not exist, the tagging fails.

If the workspace is not on the Jenkins controller, the tagging fails.

Even if the workspace exists and is on the controller, the tag is not
pushed to any other repository.

Require administrators to intentionally add the git tag action from
their Jenkins installation.  There is no benefit to cluttering Jenkins
jobs with an action that most administrators and most users will not use.